feat(passkey): separate modal for user passkey naming
